Train/SHMTU_CAS_OCR_RESNET/requirements.txt
opencv-python>=4.5.0
numpy<2.0.0
scikit-learn
pillow

torch>=2.0.0
torchvision
tqdm

selenium>=4.0.0
webdriver-manager

matplotlib

jupyter
notebook

openvino-dev

onnx
onnxruntime
onnxconverter-common
